refactor(stellar): unify getExtra and getSigners tests into facilitator-accessors
Merge facilitator-getExtra.test.ts and facilitator-getSigners.test.ts
into a single facilitator-accessors.test.ts, eliminating duplicate
imports, mocks, and setup.
